Redefine pixel to use unsigned 8-bit integers

+++ canvas.h
@@ -5,9 +5,9 @@
 
 
 typedef struct {
-    int red;
-    int green;
-    int blue;
+    uint8_t red;
+    uint8_t green;
+    uint8_t blue;
 } pixel;
 
 typedef struct {
